Restored auto reconnect and printing invocation boundaries
